% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanmem.R
\name{cleanMem}
\alias{cleanMem}
\title{Garbage collection function ruthlessly copied from https://stackoverflow.com/a/1467334}
\usage{
cleanMem(n = 10)
}
\arguments{
\item{n}{number of gc() iterations}
}
\description{
Garbage collection function ruthlessly copied from https://stackoverflow.com/a/1467334
}
