% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cleanup}
\alias{cleanup}
\title{Remove objects from the global environment}
\usage{
cleanup(exclude = c())
}
\arguments{
\item{exclude}{which objects not to remove}
}
\description{
Remove objects from the global environment
}
