% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.add_data_type}
\alias{.add_data_type}
\title{Builds the file paths for one or several file types}
\usage{
.add_data_type(file_dir, type)
}
\arguments{
\item{file_dir}{file path without or with file type ending}

\item{type}{one or several file types}
}
\value{
file paths, one per given type
}
\description{
Builds the file paths for one or several file types
}
\keyword{internal}
