% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_data.R
\name{.cat_character_named}
\alias{.cat_character_named}
\title{Prints vectors to console in as character vector}
\usage{
.cat_character_named(...)
}
\arguments{
\item{...}{vector/s}
}
\description{
Prints vectors to console in as character vector
}
