% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.check_data_type}
\alias{.check_data_type}
\title{Checks file types given to write_data()}
\usage{
.check_data_type(type)
}
\arguments{
\item{type}{one or several file types}
}
\value{
lowercase vector of unique, valid file types
}
\description{
Checks file types given to write_data()
}
\keyword{internal}
