% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.data_type_delim}
\alias{.data_type_delim}
\title{Delimiter used for a delimited file type}
\usage{
.data_type_delim(type)
}
\arguments{
\item{type}{file type ("tsv", "csv" or "txt")}
}
\value{
single character delimiter (NA for non-delimited types)
}
\description{
Delimiter used for a delimited file type
}
\keyword{internal}
