% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.save_objects_recursively}
\alias{.save_objects_recursively}
\title{Saves a nested list as a folder of single files}
\usage{
.save_objects_recursively(
  object,
  name,
  dir,
  type = NULL,
  silent = F,
  redo = T,
  list_as_folders = T,
  clean_memory = F,
  partitioning = NULL,
  ...
)
}
\arguments{
\item{object}{list or single object to be saved}

\item{name}{name of the folder or file}

\item{dir}{folder the object is saved in}

\item{type}{(optional) file type used for every table in the list}

\item{silent}{Should messages be suppressed?}

\item{redo}{Should existing files be written again?}

\item{list_as_folders}{Should nested lists become nested folders?}

\item{clean_memory}{Should the memory be cleaned with gc()/cleanMem() afer
writing}

\item{partitioning}{Should the parquet files be split into}

\item{...}{additional arguments for the saving function}
}
\value{
TRUE (invisibly)
}
\description{
Saves a nested list as a folder of single files
}
