% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations.R
\name{.save_observations_data}
\alias{.save_observations_data}
\title{Title}
\usage{
.save_observations_data(
  observations_data_frame,
  observations_data_frame_preview,
  name = "observations",
  tag = "",
  dataset,
  n_preview = 100,
  save_dir,
  silent = F
)
}
\arguments{
\item{observations_data_frame}{}

\item{observations_data_frame_preview}{}

\item{name}{}

\item{tag}{}

\item{dataset}{}

\item{n_preview}{}

\item{save_dir}{}

\item{silent}{}
}
\description{
Title
}
