% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables.R
\name{.save_variables_data}
\alias{.save_variables_data}
\title{Title}
\usage{
.save_variables_data(
  variables_data_frame,
  variables_data_frame_preview,
  dataset,
  name = "variables",
  tag = "",
  n_preview = 100,
  save_dir
)
}
\arguments{
\item{variables_data_frame}{}

\item{variables_data_frame_preview}{}

\item{dataset}{}

\item{name}{}

\item{tag}{}

\item{n_preview}{}

\item{save_dir}{}
}
\description{
Title
}
