% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.trim_data_type}
\alias{.trim_data_type}
\title{Removes a known file type ending from a file path}
\usage{
.trim_data_type(file_dir)
}
\arguments{
\item{file_dir}{file path}
}
\value{
file path without a trailing known file type ending
}
\description{
Removes a known file type ending from a file path
}
\keyword{internal}
