% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.write_data_file}
\alias{.write_data_file}
\title{Writes a single file of a given file type}
\usage{
.write_data_file(x, file_dir, type, silent = F, partitioning = NULL, ...)
}
\arguments{
\item{x}{data to be saved}

\item{file_dir}{final file path including the file type ending}

\item{type}{file type to write}

\item{silent}{Should messages be suppressed?}

\item{partitioning}{partitioning columns (parquet only)}

\item{...}{additional arguments for the saving function}
}
\value{
file path of the written file (invisibly)
}
\description{
Writes a single file of a given file type
}
\keyword{internal}
