% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fus}
\alias{fus}
\title{F(actors) U(nique) S(orted)}
\usage{
fus(x, decreasing = F)
}
\arguments{
\item{x}{a vector}

\item{decreasing}{sort in decreasing order (default=F)}
}
\description{
F(actors) U(nique) S(orted)
}
