% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00d_data_frames.R
\name{get_data_frame}
\alias{get_data_frame}
\title{Assemble data frame from dataset}
\usage{
get_data_frame(
  which,
  variables,
  observations,
  dataset,
  output.type = "tibble",
  ...
)
}
\arguments{
\item{which}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{dataset}{dataset name}

\item{output.type}{data type (default = "tibble", "data.frame", "matrix")}

\item{...}{additional arguments for open_data()}
}
\description{
Assemble data frame from dataset
}
