% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{get_data_m}
\alias{get_data_m}
\title{Combines a get_data(), map() and bind_rows()}
\usage{
get_data_m(file, recursive = T, credit = 0, .id = NULL)
}
\arguments{
\item{file}{file name}

\item{recursive}{Should data be recursively loaded?}

\item{credit}{how many recursive steps are allowed}

\item{.id}{name of an optional identifier column (see dplyr::bind_rows())}
}
\description{
Combines a get_data(), map() and bind_rows()
}
