% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations.R
\name{get_observations_data}
\alias{get_observations_data}
\title{Return observations data}
\usage{
get_observations_data(which, observations, dataset, as_arrow_table = F, ...)
}
\arguments{
\item{which}{which observations data columns to pull (multiple supported)}

\item{observations}{(optional) vector of observations or expression}

\item{dataset}{dataset}

\item{as_arrow_table}{return an arrow table}

\item{...}{additional arguments to open_data()}
}
\description{
Return observations data
}
