% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables.R
\name{get_variables}
\alias{get_variables}
\title{Return variables}
\usage{
get_variables(variables, dataset, ...)
}
\arguments{
\item{variables}{vector of variables}

\item{dataset}{dataset}

\item{...}{additional arguments to open_data()}
}
\description{
Return variables
}
