% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables.R
\name{get_variables_data}
\alias{get_variables_data}
\title{Return variables data}
\usage{
get_variables_data(which, variables, dataset, as_arrow_table = F, ...)
}
\arguments{
\item{which}{which variables data columns to pull (multiple supported)}

\item{variables}{(optional) vector of variables or expression}

\item{dataset}{dataset}

\item{as_arrow_table}{return an arrow table}

\item{...}{additional arguments to open_data()}
}
\description{
Return variables data
}
