% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_diann.R
\name{import_diann_channel}
\alias{import_diann_channel}
\title{Title}
\usage{
import_diann_channel(
  file = "report.parquet",
  name = "Precursors",
  channel = "Channel",
  filter_by = Proteotypic == 1 & Decoy == 0,
  observation_names = list(pattern = ".+"),
  variables_data = "default",
  data_frames = "default",
  preview_precursors = c("all", "top100"),
  preview_format = c("wide_obs", "wide_vars", "long"),
  save_dir,
  partition_by_run = F,
  silent = F
)
}
\arguments{
\item{file}{}

\item{name}{}

\item{channel}{}

\item{filter_by}{}

\item{observation_names}{}

\item{variables_data}{}

\item{data_frames}{}

\item{preview_precursors}{}

\item{preview_format}{}

\item{save_dir}{}

\item{partition_by_run}{}
}
\description{
Title
}
