% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{load_objects}
\alias{load_objects}
\title{Title}
\usage{
load_objects(
  dir = "",
  objects = c("Analysis", "Datasets", "Info"),
  exclude = NULL,
  assign = T,
  silent = F
)
}
\arguments{
\item{dir}{folder or folders to read}

\item{objects}{names of the top-level objects to read}

\item{exclude}{pattern of entries to return as a path instead of reading}

\item{assign}{assign the objects into the global environment}

\item{silent}{Should messages be suppressed?}
}
\value{
the list of read objects (invisibly)
}
\description{
Title
}
