% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{open_data}
\alias{open_data}
\title{Loads data from a file name}
\usage{
open_data(file, fallback, recursive = T, credit = 10, ...)
}
\arguments{
\item{file}{file name}

\item{fallback}{other file name or alternative way to provide the input -
useful if file is an R object and fallback is a hardcoded string}

\item{recursive}{Should data be recursively loaded?}

\item{credit}{how many recursive steps are allowed}

\item{...}{additional arguments for the opening function}
}
\value{
an Arrow dataset connection, or the object itself if \if{html}{\out{<file>}} is not
a file path
}
\description{
Loads data from a file name
}
\examples{
  data_small <- tibble::tibble(a = 1:3, b = letters[1:3]) \%>\%
    write_data(file = "data_small", 
               dir = tempdir(), 
               type = "tsv")

  open_data(data_small) \%>\%
    dplyr::filter(a > 1) \%>\%
    dplyr::collect()
}
