% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00a_datasets.R
\name{retag_datasets}
\alias{retag_datasets}
\title{Change name or copy data before saving the RData environment}
\usage{
retag_datasets(tag = "_x", copy.files = FALSE)
}
\arguments{
\item{tag}{suffix for data names}

\item{copy.files}{should data frames with other names be copied (default = F)}
}
\description{
Change name or copy data before saving the RData environment
}
