% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00d_data_frames.R
\name{save_data_frame}
\alias{save_data_frame}
\title{Title}
\usage{
save_data_frame(
  data,
  dataset,
  name = "data",
  tag = "",
  n_preview = 100,
  save_dir,
  partitioning
)
}
\arguments{
\item{data}{}

\item{dataset}{}

\item{name}{}

\item{tag}{}

\item{n_preview}{}

\item{save_dir}{}

\item{partitioning}{}
}
\description{
Title
}
