% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{save_objects}
\alias{save_objects}
\title{Title}
\usage{
save_objects(
  objects = c("Analysis", "Datasets", "Info"),
  dir = "Data/RData/temp",
  silent = F
)
}
\arguments{
\item{objects}{}

\item{dir}{}

\item{silent}{}
}
\description{
Title
}
