% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00c_observations.R
\name{save_observations_data}
\alias{save_observations_data}
\title{Title}
\usage{
save_observations_data(
  data,
  columns,
  dataset,
  name = "observations",
  tag = "",
  n_preview = 100,
  save_dir
)
}
\arguments{
\item{data}{observations data as tibble or Arrow object}

\item{columns}{}

\item{dataset}{dataset}

\item{name}{}

\item{tag}{}

\item{n_preview}{}

\item{save_dir}{}
}
\description{
Title
}
