% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00b_variables.R
\name{save_variables_data}
\alias{save_variables_data}
\title{Title}
\usage{
save_variables_data(
  data,
  columns,
  name = "variables",
  tag = "",
  dataset,
  n_preview = 100,
  save_dir
)
}
\arguments{
\item{data}{variables data as tibble or Arrow object}

\item{columns}{}

\item{name}{}

\item{tag}{}

\item{dataset}{}

\item{n_preview}{}

\item{save_dir}{}
}
\description{
Title
}
