% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{tempdir_remove}
\alias{tempdir_remove}
\title{Removing all temporary files saved by write_data() when}
\usage{
tempdir_remove(
  dir = tempdir(),
  all.paths = F,
  pattern = ".Rds|.parquet|.tsv|.csv|.txt|.pdf"
)
}
\arguments{
\item{dir}{location of temporary files}

\item{all.paths}{find all temporary folders}

\item{pattern}{pattern/s for files to be removed}
}
\description{
Removing all temporary files saved by write_data() when
}
\examples{
  tempdir_remove()
}
