% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{tempdir_size}
\alias{tempdir_size}
\title{List size of all temporary files saved by write_data() when}
\usage{
tempdir_size(
  dir = tempdir(),
  all.paths = F,
  pattern = ".Rds|.parquet|.tsv|.csv|.txt",
  units = "auto_si"
)
}
\arguments{
\item{dir}{location of temporary files}

\item{all.paths}{find all temporary folders}

\item{pattern}{pattern/s for files to be removed}

\item{units}{unit/s to use to represent file size}
}
\description{
List size of all temporary files saved by write_data() when
}
\examples{
  tempdir_list()
}
