% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_data.R
\name{create_preview}
\alias{create_preview}
\title{Substitutes the given data frame with a preview, saves the data locally and
stores the location in a hidden variable}
\usage{
create_preview(x, n = 10, clean_memory = T)
}
\description{
Substitutes the given data frame with a preview, saves the data locally and
stores the location in a hidden variable
}
