% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_data.R
\name{.preview_one}
\alias{.preview_one}
\title{Collects the first rows of one opened object}
\usage{
.preview_one(x, n)
}
\arguments{
\item{x}{an opened connection or a data frame}

\item{n}{number of rows to read}
}
\value{
a tibble of the first \if{html}{\out{<n>}} rows carrying a \code{total_rows} attribute
}
\description{
Collects the first rows of one opened object
}
\keyword{internal}
