% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.read_objects_recursively}
\alias{.read_objects_recursively}
\title{Title}
\usage{
.read_objects_recursively(
  name,
  dir,
  exclude = NULL,
  silent = F,
  as_arrow_table = F,
  ...
)
}
\arguments{
\item{name}{name of the folder or file to read}

\item{dir}{folder the object lives in}

\item{exclude}{pattern of entries to return as a path instead of reading}

\item{silent}{Should messages be suppressed?}

\item{as_arrow_table}{return tibbles or Arrow connections}

\item{...}{additional arguments for the reading function}
}
\value{
the nested list rebuilt from the folder tree
}
\description{
Title
}
