% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.read_objects_recursively}
\alias{.read_objects_recursively}
\title{Title}
\usage{
.read_objects_recursively(name, dir, exclude = NULL, silent = F)
}
\arguments{
\item{name}{}

\item{dir}{}

\item{exclude}{}

\item{silent}{}
}
\description{
Title
}
