% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00d_data_frames.R
\name{.save_data_frame}
\alias{.save_data_frame}
\title{Title}
\usage{
.save_data_frame(
  data_frame,
  data_frame_preview,
  dataset,
  name = "file",
  tag = "",
  n_preview = 100,
  save_dir,
  partitioning,
  silent = F
)
}
\arguments{
\item{data_frame}{}

\item{data_frame_preview}{}

\item{dataset}{}

\item{name}{}

\item{tag}{}

\item{n_preview}{}

\item{save_dir}{}

\item{partitioning}{}

\item{silent}{}
}
\description{
Title
}
