% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.save_objects_recursively}
\alias{.save_objects_recursively}
\title{Title}
\usage{
.save_objects_recursively(
  object,
  name,
  dir,
  silent = F,
  redo = T,
  list_as_folders = T,
  clean_memory = F,
  partitioning = NULL,
  ...
)
}
\arguments{
\item{object}{}

\item{name}{}

\item{dir}{}

\item{silent}{}

\item{redo}{}

\item{list_as_folders}{}

\item{clean_memory}{}

\item{partitioning}{}

\item{...}{}
}
\description{
Title
}
