% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{.save_objects_recursively}
\alias{.save_objects_recursively}
\title{Title}
\usage{
.save_objects_recursively(object, name, dir, silent = F)
}
\arguments{
\item{object}{}

\item{name}{}

\item{dir}{}

\item{silent}{}
}
\description{
Title
}
