% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{get_data}
\alias{get_data}
\title{Loads data from a file name}
\usage{
get_data(file, recursive = T, credit = 10, as_arrow_table = F, ...)
}
\arguments{
\item{file}{file name}

\item{recursive}{Should data be recursively loaded?}

\item{credit}{how many recursive steps are allowed}

\item{as_arrow_table}{return a tibble or an Arrow connection}
}
\description{
Loads data from a file name
}
