% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{load_objects}
\alias{load_objects}
\title{Title}
\usage{
load_objects(
  dir = "",
  objects = c("Analysis", "Datasets", "Info"),
  exclude = NULL,
  assign = T,
  silent = F
)
}
\arguments{
\item{dir}{}

\item{objects}{}

\item{exclude}{}

\item{silent}{}
}
\description{
Title
}
