% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{open_data}
\alias{open_data}
\title{Loads data from a file name}
\usage{
open_data(file, recursive = T, credit = 10, ...)
}
\arguments{
\item{file}{file name}

\item{recursive}{Should data be recursively loaded?}

\item{credit}{how many recursive steps are allowed}
}
\description{
Loads data from a file name
}
