% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{tempdir_list}
\alias{tempdir_list}
\title{List all temporary files saved by write_data() when}
\usage{
tempdir_list(dir = tempdir(), pattern = ".Rds|.parquet")
}
\arguments{
\item{dir}{location of temporary files}

\item{pattern}{pattern/s for files to be removed}
}
\description{
List all temporary files saved by write_data() when
}
\examples{
  tempdir_list()
}
