% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_write_data.R
\name{write_data}
\alias{write_data}
\title{Title}
\usage{
write_data(
  x,
  file,
  dir,
  dir.create = F,
  clean_memory = F,
  silent = T,
  partitioning
)
}
\arguments{
\item{x}{}

\item{file}{}

\item{dir}{}

\item{dir.create}{}

\item{clean_memory}{}

\item{silent}{}

\item{partitioning}{}
}
\description{
Title
}
